Clean up Circe section
