+++ src/main/scala/com/reagroup/exercises/circe/README.md
@@ -1,0 +1,15 @@
+# Circe Exercises
+
+Circe (pronounced SUR-see, or KEER-kee in classical Greek, or CHEER-chay in Ecclesiastical Latin) is a JSON library for Scala.
+
+We like Circe as opposed to other libraries because it is functional, type-safe and very idiomatic.
+It integrates very well with the Cats ecosystem.
+
+For more comprehensive docs on Circe:
+https://circe.github.io/circe/
+
+There are 3 parts to these exercises. The recommended progression is:
+
+1. Parsing (`String => Json`)
+1. Decoding (`Json => A`)
+1. Encoding (`A => Json`)
