--- src/main/scala/com/reagroup/exercises/circe/a/CirceParsingExercises.scala
@@ -1,47 +1,0 @@
-package com.reagroup.exercises.circe.a
-
-
-import io.circe._
-
-/**
-  * Circe (pronounced SUR-see, or KEER-kee in classical Greek, or CHEER-chay in Ecclesiastical Latin) is a JSON library for Scala.
-  *
-  * We like Circe as opposed to other libraries because it is functional, type-safe and very idiomatic.
-  * It integrates very well with the Cats ecosystem.
-  *
-  * For more comprehensive docs on Circe:
-  * https://circe.github.io/circe/
-  *
-  * There are 3 parts to these exercises.
-  *
-  * 1. Parsing (`String => Json`)
-  * 2. Encoding (`A => Json`)
-  * 3. Decoding (`Json => A`)
-  */
-object CirceParsingExercises {
-  /**
-    * Json Parsing
-    *
-    * Hint: `parser` is already in scope (imported through `io.circe._`)
-    *
-    * Why is the return type an `Either`?
-    */
-  def strToJson(str: String): Either[ParsingFailure, Json] = {
-    ???
-  }
-
-  /**
-    * Try make a syntax error in the following Json document and compile.
-    * What happens?
-    */
-  val validJson: Json = {
-    import io.circe.literal._
-
-    json"""
-      {
-        "someKey": "someValue",
-        "anotherKey": "anotherValue"
-      }
-    """
-  }
-}
